backend/pyproject.toml
[project]
name = "autoshorts-backend"
version = "0.1.0"
description = "AutoShorts MVP - Automatyczna generacja faceless short-video"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.2.4",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.27.0",
    "celery[redis]>=5.3.6",
    "redis>=5.0.1",
    "openai>=1.12.0",
    "elevenlabs>=0.2.0",
    "boto3>=1.34.0",
    "stripe>=8.0.0",
    "google-api-python-client>=2.116.0",
    "google-auth-oauthlib>=1.2.0",
    "ffmpeg-python>=0.2.0",
    "Pillow>=10.2.0",
    "numpy>=1.26.0",
    "authlib>=1.3.0",
    "slowapi>=0.1.9",
    "sentry-sdk[fastapi]>=1.40.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.3",
    "aiofiles>=23.2.0",
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S101", "B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
